Extracted common code to abstract class
