+++ ihmc-high-level-behaviors/src/libgdx/java/us/ihmc/rdx/ui/graphics/RDXRawImagePointCloudRenderer.java
@@ -1,53 +1,35 @@
 package us.ihmc.rdx.ui.graphics;
 
-import com.badlogic.gdx.graphics.Color;
-import com.badlogic.gdx.graphics.Mesh;
 import com.badlogic.gdx.graphics.Pixmap;
 import com.badlogic.gdx.graphics.Pixmap.Format;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.VertexAttribute;
+import com.badlogic.gdx.graphics.VertexAttributes;
 import com.badlogic.gdx.graphics.VertexAttributes.Usage;
-import com.badlogic.gdx.graphics.g3d.Material;
-import com.badlogic.gdx.graphics.g3d.Renderable;
-import com.badlogic.gdx.graphics.g3d.RenderableProvider;
-import com.badlogic.gdx.graphics.g3d.shaders.DefaultShader;
 import com.badlogic.gdx.graphics.glutils.PixmapTextureData;
 import com.badlogic.gdx.math.Matrix4;
-import com.badlogic.gdx.utils.Array;
-import com.badlogic.gdx.utils.Pool;
-import net.mgsx.gltf.scene3d.attributes.PBRColorAttribute;
 import org.bytedeco.javacpp.BytePointer;
 import org.bytedeco.javacpp.FloatPointer;
 import org.bytedeco.opencv.global.opencv_core;
 import org.bytedeco.opencv.opencv_core.Mat;
-import org.lwjgl.opengl.GL41;
+import org.jetbrains.annotations.Nullable;
 import us.ihmc.euclid.referenceFrame.FramePose3D;
 import us.ihmc.euclid.referenceFrame.interfaces.FixedFramePose3DBasics;
 import us.ihmc.euclid.transform.RigidBodyTransform;
 import us.ihmc.perception.RawImage;
 import us.ihmc.perception.camera.CameraIntrinsics;
+import us.ihmc.rdx.AbstractRDXPointCloudRenderer;
 import us.ihmc.rdx.shader.RDXShader;
 import us.ihmc.rdx.shader.RDXUniform;
 import us.ihmc.rdx.tools.LibGDXTools;
 
 import java.nio.FloatBuffer;
 
-public class RDXRawImagePointCloudRenderer implements RenderableProvider
+import static us.ihmc.rdx.AbstractRDXPointCloudRenderer.ColoringMethod.*;
+
+public class RDXRawImagePointCloudRenderer extends AbstractRDXPointCloudRenderer
 {
-   public enum ColoringMethod
-   {
-      DEFAULT, GRADIENT_WORLD_Z, GRADIENT_SENSOR_X, COLOR_IMAGE
-   }
-
-   private Renderable renderable;
-
-   private final VertexAttribute depthDataAttribute = new VertexAttribute(Usage.Generic, 1, "a_depthData");
-
-   // GENERALLY NEEDED UNIFORMS
-   private float pointScale = 0.01f;
-
-   private final Color defaultPointColor = new Color(Color.WHITE);
-   private ColoringMethod coloringMethod = ColoringMethod.DEFAULT;
+   private final VertexAttributes vertexAttributes = new VertexAttributes(new VertexAttribute(Usage.Generic, 1, "a_depthData"));
 
    // DEPTH IMAGE UNIFORM DATA
    private CameraIntrinsics depthIntrinsics = new CameraIntrinsics();
@@ -70,119 +52,6 @@ public class RDXRawImagePointCloudRenderer implements RenderableProvider
    public RDXRawImagePointCloudRenderer(boolean enableColorImageInput)
    {
       this.enableColorImageInput = enableColorImageInput;
-   }
-
-   public void create(int maxPoints)
-   {
-      GL41.glEnable(GL41.GL_VERTEX_PROGRAM_POINT_SIZE);
-
-      renderable = new Renderable();
-      renderable.meshPart.primitiveType = GL41.GL_POINTS;
-      renderable.meshPart.offset = 0;
-      renderable.material = new Material(PBRColorAttribute.createBaseColorFactor(Color.WHITE));
-
-      boolean isStatic = false;
-      int maxIndices = 0;
-      renderable.meshPart.mesh = new Mesh(isStatic, maxPoints, maxIndices, depthDataAttribute);
-
-      String[] vertexFlags = { "INPUT_COLOR_IMAGE" };
-      if (!enableColorImageInput)
-         vertexFlags = null;
-
-      RDXShader shader = new RDXShader(getClass());
-      shader.create(vertexFlags, null);
-      registerGeneralUniforms(shader);
-      registerDepthImageUniforms(shader);
-
-      if (enableColorImageInput)
-         registerColorImageUniforms(shader);
-
-      shader.init(renderable);
-      renderable.shader = shader.getBaseShader();
-   }
-
-   private void registerGeneralUniforms(RDXShader rdxShader)
-   {
-      rdxShader.getBaseShader().register(DefaultShader.Inputs.viewTrans, DefaultShader.Setters.viewTrans);
-      rdxShader.getBaseShader().register(DefaultShader.Inputs.projTrans, DefaultShader.Setters.projTrans);
-
-      RDXUniform screenWidthUniform = RDXUniform.createGlobalUniform("u_screenWidth", (shader, inputID, renderable, combinedAttributes) ->
-      {
-         shader.set(inputID, shader.camera.viewportWidth);
-      });
-      rdxShader.registerUniform(screenWidthUniform);
-
-      RDXUniform pointScaleUniform = RDXUniform.createGlobalUniform("u_pointScale", (shader, inputID, renderable, combinedAttributes) ->
-      {
-         shader.set(inputID, pointScale);
-      });
-      rdxShader.registerUniform(pointScaleUniform);
-
-      RDXUniform defaultPointColorUniform = RDXUniform.createGlobalUniform("u_defaultPointColor", (shader, inputID, renderable, combinedAttributes) ->
-      {
-         shader.set(inputID, defaultPointColor);
-      });
-      rdxShader.registerUniform(defaultPointColorUniform);
-
-      RDXUniform coloringMethodUniform = RDXUniform.createGlobalUniform("u_coloringMethod", (shader, inputID, renderable, combinedAttributes) ->
-      {
-         shader.set(inputID, coloringMethod.ordinal());
-      });
-
-      rdxShader.registerUniform(coloringMethodUniform);
-   }
-
-   private void registerDepthImageUniforms(RDXShader rdxShader)
-   {
-      RDXUniform depthIntrinsicsUniform = RDXUniform.createGlobalUniform("u_depthIntrinsics", (shader, inputID, renderable, combinedAttributes) ->
-      {
-         shader.set(inputID, (float) depthIntrinsics.getFx(), (float) depthIntrinsics.getFy(), (float) depthIntrinsics.getCx(), (float) depthIntrinsics.getCy());
-      });
-      rdxShader.registerUniform(depthIntrinsicsUniform);
-
-      RDXUniform depthImageWidthUniform = RDXUniform.createGlobalUniform("u_depthImageWidth", (shader, inputID, renderable, combinedAttributes) ->
-      {
-         shader.set(inputID, depthIntrinsics.getWidth());
-      });
-      rdxShader.registerUniform(depthImageWidthUniform);
-
-      RDXUniform depthDiscretizationUniform = RDXUniform.createGlobalUniform("u_depthDiscretization", (shader, inputID, renderable, combinedAttributes) ->
-      {
-         shader.set(inputID, depthDiscretization);
-      });
-      rdxShader.registerUniform(depthDiscretizationUniform);
-
-      RDXUniform depthTransformUniform = RDXUniform.createGlobalUniform("u_depthTransform", (shader, inputID, renderable, combinedAttributes) ->
-      {
-         shader.set(inputID, libGDXDepthTransform);
-      });
-      rdxShader.registerUniform(depthTransformUniform);
-   }
-
-   private void registerColorImageUniforms(RDXShader rdxShader)
-   {
-      colorImagePixmap = new Pixmap(1, 1, Format.RGBA8888);
-      pixmapDataPointer = new BytePointer(colorImagePixmap.getPixels());
-      colorImageTexture = new Texture(colorImagePixmap);
-
-      RDXUniform colorImageTextureUniform = RDXUniform.createGlobalUniform("u_colorTexture", (shader, inputID, renderable, combinedAttributes) ->
-      {
-         shader.set(inputID, colorImageTexture);
-      });
-      rdxShader.registerUniform(colorImageTextureUniform);
-
-      RDXUniform colorIntrinsicsUniform = RDXUniform.createGlobalUniform("u_colorIntrinsics", (shader, inputID, renderable, combinedAttributes) ->
-      {
-         shader.set(inputID, (float) colorIntrinsics.getFx(), (float) colorIntrinsics.getFy(), (float) colorIntrinsics.getCx(), (float) colorIntrinsics.getCy());
-      });
-      rdxShader.registerUniform(colorIntrinsicsUniform);
-
-      RDXUniform depthTransformUniform = RDXUniform.createGlobalUniform("u_depthToColorTransform", (shader, inputID, renderable, combinedAttributes) ->
-      {
-         LibGDXTools.toLibGDX(depthToColorTransform, tempColorTransform, libGDXColorTransform);
-         shader.set(inputID, libGDXColorTransform);
-      });
-      rdxShader.registerUniform(depthTransformUniform);
    }
 
    public void updateMesh(RawImage depthImage)
@@ -262,30 +131,97 @@ public class RDXRawImagePointCloudRenderer implements RenderableProvider
    }
 
    @Override
-   public void getRenderables(Array<Renderable> renderables, Pool<Renderable> pool)
+   public ColoringMethod[] getAvailableColoringMethods()
    {
-      if (renderable != null)
-         renderables.add(renderable);
+      if (enableColorImageInput)
+         return new ColoringMethod[] {DEFAULT, GRADIENT_WORLD_Z, GRADIENT_SENSOR_X, COLOR_IMAGE};
+
+      return new ColoringMethod[] {DEFAULT, GRADIENT_WORLD_Z, GRADIENT_SENSOR_X};
    }
 
-   public void dispose()
+   @Nullable
+   @Override
+   protected String[] getFragmentShaderFlags()
    {
-      if (renderable.meshPart.mesh != null)
-         renderable.meshPart.mesh.dispose();
+      return null;
    }
 
-   public void setPointScale(float size)
+   @Nullable
+   @Override
+   protected String[] getVertexShaderFlags()
    {
-      pointScale = size;
+      if (enableColorImageInput)
+         return new String[] {"INPUT_COLOR_IMAGE"};
+
+      return null;
    }
 
-   public void setDefaultPointColor(Color color)
+   @Override
+   protected VertexAttributes getVertexAttributes()
    {
-      defaultPointColor.set(color);
+      return vertexAttributes;
    }
 
-   public void setColoringMethod(ColoringMethod method)
+   @Override
+   protected void registerUniforms(RDXShader rdxShader)
    {
-      coloringMethod = method;
+      registerGeneralUniforms(rdxShader);
+      registerDepthImageUniforms(rdxShader);
+
+      if (enableColorImageInput)
+         registerColorImageUniforms(rdxShader);
+   }
+
+   private void registerDepthImageUniforms(RDXShader rdxShader)
+   {
+      RDXUniform depthIntrinsicsUniform = RDXUniform.createGlobalUniform("u_depthIntrinsics", (shader, inputID, renderable, combinedAttributes) ->
+      {
+         shader.set(inputID, (float) depthIntrinsics.getFx(), (float) depthIntrinsics.getFy(), (float) depthIntrinsics.getCx(), (float) depthIntrinsics.getCy());
+      });
+      rdxShader.registerUniform(depthIntrinsicsUniform);
+
+      RDXUniform depthImageWidthUniform = RDXUniform.createGlobalUniform("u_depthImageWidth", (shader, inputID, renderable, combinedAttributes) ->
+      {
+         shader.set(inputID, depthIntrinsics.getWidth());
+      });
+      rdxShader.registerUniform(depthImageWidthUniform);
+
+      RDXUniform depthDiscretizationUniform = RDXUniform.createGlobalUniform("u_depthDiscretization", (shader, inputID, renderable, combinedAttributes) ->
+      {
+         shader.set(inputID, depthDiscretization);
+      });
+      rdxShader.registerUniform(depthDiscretizationUniform);
+
+      RDXUniform depthTransformUniform = RDXUniform.createGlobalUniform("u_depthTransform", (shader, inputID, renderable, combinedAttributes) ->
+      {
+         shader.set(inputID, libGDXDepthTransform);
+      });
+      rdxShader.registerUniform(depthTransformUniform);
+   }
+
+   private void registerColorImageUniforms(RDXShader rdxShader)
+   {
+      colorImagePixmap = new Pixmap(1, 1, Format.RGBA8888);
+      pixmapDataPointer = new BytePointer(colorImagePixmap.getPixels());
+      colorImageTexture = new Texture(colorImagePixmap);
+
+      RDXUniform colorImageTextureUniform = RDXUniform.createGlobalUniform("u_colorTexture", (shader, inputID, renderable, combinedAttributes) ->
+      {
+         shader.set(inputID, colorImageTexture);
+      });
+      rdxShader.registerUniform(colorImageTextureUniform);
+
+      RDXUniform colorIntrinsicsUniform = RDXUniform.createGlobalUniform("u_colorIntrinsics", (shader, inputID, renderable, combinedAttributes) ->
+      {
+         shader.set(inputID, (float) colorIntrinsics.getFx(), (float) colorIntrinsics.getFy(), (float) colorIntrinsics.getCx(), (float) colorIntrinsics.getCy());
+      });
+      rdxShader.registerUniform(colorIntrinsicsUniform);
+
+      RDXUniform depthTransformUniform = RDXUniform.createGlobalUniform("u_depthToColorTransform", (shader, inputID, renderable, combinedAttributes) ->
+      {
+         LibGDXTools.toLibGDX(depthToColorTransform, tempColorTransform, libGDXColorTransform);
+         shader.set(inputID, libGDXColorTransform);
+      });
+      rdxShader.registerUniform(depthTransformUniform);
    }
 }
